refactor: 수수료 계산은 OrderSettlement 엔티티에서 사용

+++ batch/src/main/java/store/_0982/batch/batch/sellerpayout/writer/RetryFailedSellerPayoutWriter.java
@@ -52,7 +52,7 @@ public class RetryFailedSellerPayoutWriter implements ItemWriter<SellerPayout> {
                 }
 
                 sellerPayout.setAccountInfo(accountInfo.accountNumber(), accountInfo.bankCode());
-                bankTransferService.transfer(accountInfo, sellerPayout.getTotalAmount().longValue());
+                bankTransferService.transfer(accountInfo, sellerPayout.getTotalAmount());
                 sellerPayout.markAsCompleted();
                 sellerBalanceService.clearBalance(sellerPayout);
